refactor(hooks): decompose useEasyModeView to resolve max-lines-per-function ESLint rule
